epp/web/epp/requirements.txt
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
asgiref==3.2.3
Django==3.0.3
Pillow==7.0.0
pip==20.0.2
pytz==2019.3
setuptools==40.6.2
sqlparse==0.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.11.
google-cloud-vision==0.42.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.8
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
django-chartjs==2.0.0
django-crispy-forms==1.9.0